requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
facebook-sdk==3.1.0
tqdm==4.66.1
spacy==3.7.2